+++ ChangeLog.md
@@ -9,6 +9,7 @@
 * Display information on why a snapshot was rejected [#381](https://github.com/commercialhaskell/stack/issues/381)
 * Give a reason for unregistering packages [#389](https://github.com/commercialhaskell/stack/issues/389)
 * `stack exec` accepts the `--no-ghc-package-path` parameter
+* Don't require build plan to upload [#400](https://github.com/commercialhaskell/stack/issues/400)
 
 Bug fixes: